Rename files reward to dynamicReward
