+++ services/blockchain-connector/methods/dynamicReward.js
