+++ services/blockchain-connector/shared/sdk/index.js
@@ -76,7 +76,7 @@ const {
 const {
 	getRewardTokenID,
 	getDefaultRewardAtHeight,
-} = require('./reward');
+} = require('./dynamicReward');
 
 const {
 	getFeeTokenID,
